Change the underlying type of raven floats from long double to double. Add int and float hashing support for the hash object. Add some guard code to switch statements in debug module. Add a check to terminate the repl on EOF (^D), not just interrupt (^C).

+++ src/ast.h
@@ -184,7 +184,7 @@ struct AST_patt {
         AST_pair_patt pair;
         char *ident;       /* identifier(variable) pattern */
         int64_t i;         /* literal int pattern */
-        long double f;     /* literal float pattern */
+        double f;          /* literal float pattern */
         char *s;           /* literal string pattern */
         int8_t b;          /* literal boolean value */
     };
@@ -278,7 +278,7 @@ struct AST_lit_expr {
         AST_fn_lit fn;
         AST_hash_lit hash;
         AST_list_lit list;
-        long double f; /* float literal */
+        double f;      /* float literal */
         int64_t i;     /* integer literal */
         char *s;       /* string literal */
     };
